pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trellm"
version = "0.1.0"
description = "Automation tool bridging Trello boards with AI coding assistants"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "TreLLM Contributors" }
]
keywords = ["trello", "claude", "automation", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
trellm = "trellm.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["trellm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
